add incorrect only switch
trickier than expected! The issue was that questionsAll is not always populated (because we only call API
if there are no questions in the cache.) So if the incorrectOnly switch is clicked, then we need to
get all questions which caused some weirdness. solution was to use a bunch of useEffects but still
don't fully understand it tbh